% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_matrix.R
\name{average_first_years_com_mat}
\alias{average_first_years_com_mat}
\title{Average abundance of first year of community matrix}
\usage{
average_first_years_com_mat(mat = NULL, nb_sampling_to_average = NULL)
}
\description{
Average abundance of first year of community matrix
}
