% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fishing_operation.R
\name{choose_from_multiple}
\alias{choose_from_multiple}
\title{Take the sampling being the closest to the other}
\usage{
choose_from_multiple(x = NULL, quarter = NULL, mode_quarter = NULL)
}
\description{
Take the sampling being the closest to the other
}
\examples{
choose_from_multiple(x = NULL, quarter = c(1,1,2), mode_quarter = 2)
# 3
choose_from_multiple(x = c(1,3,NA), quarter = c(1,1,2), mode_quarter = 2)
choose_from_multiple(x = c(3,1,NA), quarter = c(1,1,2), mode_quarter = 2)
choose_from_multiple(x = c(NA,NA,NA), quarter = c(1,1,2), mode_quarter = 2)
choose_from_multiple(x = c(NA,NA,NA), quarter = c(1,1,2), mode_quarter = NA)
#NA
choose_from_multiple(x = c(NA,NA,NA), quarter = c(1,NA,2), mode_quarter = 2)
choose_from_multiple( x = c(1, 12.1))
choose_from_multiple(x = c(NA, 12.1))
choose_from_multiple(x = c(.09, 12.1))
}
