% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{compute_abundance_turnover_two_timesteps}
\alias{compute_abundance_turnover_two_timesteps}
\title{compute_abundance_turnover_two_timesteps}
\usage{
compute_abundance_turnover_two_timesteps(t1 = NULL, t2 = NULL)
}
\description{
Based on hillebrand et al. (2018)
}
\examples{
nsp <- 3
p1 <- tibble(species = letters[1:nsp], abundance = c(0, rep(1, nsp - 1)),
  rel_abundance = abundance / sum(abundance))
p2 <- p1
compute_abundance_turnover_two_timesteps(p1, p2)
# 0
p2$abundance[1] <- 1
p2$rel_abundance <- p2$abundance / sum(p2$abundance)
compute_abundance_turnover_two_timesteps(p1, p2)

p2$rel_abundance <- c(.5, .5, 0) 
compute_abundance_turnover_two_timesteps(p1, p2)
}
