% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{compute_codyn_turnover}
\alias{compute_codyn_turnover}
\title{Compute the turnover of species}
\usage{
compute_codyn_turnover(x = NULL, y = NULL, type = NULL)
}
\description{
Compute the turnover of species
}
\examples{
ti <- matrix(c(1,0,1,1), nrow = 2, byrow = TRUE, dimnames = list(NULL, c("pika", "sala")))
compute_codyn_turnover(x = ti[1, ], y = ti[2, ], type = "appearance")
compute_codyn_turnover(x = ti[1, ], y = ti[2, ], type = "disappearance")
ti[2,1] <- 0
compute_codyn_turnover(x = ti[1, ], y = ti[2, ], type = "appearance")
compute_codyn_turnover(x = ti[1, ], y = ti[2, ], type = "disappearance")
}
