% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{compute_dist}
\alias{compute_dist}
\title{Compute distance matrix}
\usage{
compute_dist(mat = NULL, fun = NULL, ...)
}
\arguments{
\item{mat}{community matrix}

\item{fun}{function to compute distance}
}
\description{
Compute distance matrix
}
\examples{
ti <- matrix(c(1,0,1,1), nrow = 2, byrow = TRUE, dimnames = list(NULL, c("pika", "sala")))
tir <- ti / rowSums(ti)
compute_dist(mat = tir, fun = compute_hillebrand)
compute_dist(mat = tir, fun = compute_codyn_turnover, type = "appearance")
compute_dist(mat = tir, fun = compute_codyn_turnover, type = "disappearance")
compute_dist(mat = tir, fun = compute_codyn_turnover, type = "total")
}
