% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{compute_jaccard_decomp}
\alias{compute_jaccard_decomp}
\title{Baselga decomposition of jaccard}
\usage{
compute_jaccard_decomp(x = NULL, y = NULL, type = "all")
}
\description{
Baselga decomposition of jaccard
}
\examples{

mat <- matrix(
 c(1, 0, 1, 1, 0, 1, 1, 0),
 ncol = 2,
 byrow = TRUE,
 dimnames = list(NULL, c("pika", "sala"))
)

x <- decostand(mat, method = "pa")

compute_jaccard_decomp(x = x[1, ], y = x[2, ])
compute_jaccard_decomp(x = x[1, ], y = x[3, ])
compute_jaccard_decomp(x = x[1, ], y = x[1, ])
}
