% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigal_classification_method.R
\name{compute_rigal_classif}
\alias{compute_rigal_classif}
\title{Compute rigal classification on multiple variable}
\usage{
compute_rigal_classif(
  dataset = NULL,
  y_vars = NULL,
  time_var = NULL,
  site_id_var = NULL
)
}
\arguments{
\item{dataset}{data.frame}

\item{y_vars}{character vector indicating the response variables}

\item{time_var}{character string variable describing time}

\item{site_id_var}{character string}
}
\value{
data.frame
}
\description{
Compute rigal classification on multiple variable
}
\examples{
Todo
}
