% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_match.R
\name{crop_on_buffer}
\alias{crop_on_buffer}
\title{Crop an sf object on another with a buffer}
\usage{
crop_on_buffer(to_crop = NULL, x = NULL, buffer_dist = NULL)
}
\description{
Crop an sf object on another with a buffer
}
\examples{
ti <- st_sf(
  a = 3,
  geometry = st_sfc(st_point(1:2), st_point(3:4)))
tl <- st_sf(
  a = 3,
  geometry = st_sfc(st_linestring(pts), st_linestring(pts + 12))
  )

plot(g_buf)
plot(tl, add = T)
crop_on_buffer(to_crop = tl, x = ti, buffer_dist = 12)
}
