% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fishing_operation.R
\name{filter_op}
\alias{filter_op}
\title{Selection the sampling event to keep}
\usage{
filter_op(
  op_protocol = NULL,
  selected_protocol = NULL,
  selected_abun_unit = NULL,
  nb_sampling = NULL,
  span_min = 10,
  extent_month = 1,
  convert_month_to_date = FALSE,
  return_no_filtered = FALSE
)
}
\description{
we selected:
\itemize{
\item the sampling that took place close to the median month of each site
(extent_month parameter)
\item only one sampling by site and by year, the one whose date was the closest
from median of the site
\item sites that have been monitored at least nb_sampling times
\item some protocol and abundance unit (selected_protocol, selected_abun_unit)
}
}
\details{
When precise date is absent but month and year are present, we can replace
missing date by "\%year-\%month-15" (convert_month_to_date).

NA sampling protocol and unit of abundance are removed
}
