% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_community_matrix.R
\name{get_basin_year_community_matrix}
\alias{get_basin_year_community_matrix}
\title{Get community matrix by basin and year}
\usage{
get_basin_year_community_matrix(
  x = NULL,
  loc = filtered_dataset$loc,
  min_nb_site_by_basin = NULL,
  min_year_by_basin = NULL
)
}
\description{
Get community matrix by basin and year
}
