% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity_chao_hill.R
\name{get_chao_hillnb}
\alias{get_chao_hillnb}
\title{Compute chao hillnumber coverage style}
\usage{
get_chao_hillnb(
  x = NULL,
  coverage = NULL,
  confidence_int = NULL,
  adjust_abun_density = TRUE
)
}
\description{
Compute chao hillnumber coverage style
}
\examples{

}
