% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{get_com_matrix_from_site}
\alias{get_com_matrix_from_site}
\title{replace NA by 0 in abundance data}
\usage{
get_com_matrix_from_site(
  dataset = NULL,
  y_var = NULL,
  species_var = "species",
  time_var = "year",
  long_format = TRUE
)
}
\arguments{
\item{dataset}{data.frame}

\item{y_var}{chr}

\item{species_var}{chr}

\item{time_var}{chr}

\item{long_format}{lgl}
}
\description{
replace NA by 0 in abundance data
}
\examples{
tar_load(toy_dataset)
ti <- toy_dataset \%>\%
 filter(siteid == unique(toy_dataset$siteid)[2])
get_com_matrix_from_site(dataset = ti, y_var = "abundance")
Add 0 to non-observed population in a site

get_com_matrix_from_site(dataset = ti, y_var = "abundance")
}
