% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{get_filtered_dataset}
\alias{get_filtered_dataset}
\title{Select protocol and site}
\usage{
get_filtered_dataset(
  op_protocol = NULL,
  type = NULL,
  measurement = NULL,
  site_desc_loc = NULL,
  add_var_from_protocol = NULL
)
}
\arguments{
\item{type}{character string of length one "all", "site_quali", sit}
}
\description{
Select protocol and site
}
