% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rigal_classification_method.R
\name{get_rigal_trajectory_classification}
\alias{get_rigal_trajectory_classification}
\title{Get rigal trajectory for multiple sites}
\usage{
get_rigal_trajectory_classification(
  dataset = NULL,
  y_var = NULL,
  x_var = NULL,
  site_id = NULL,
  backtransform_linear_coefs = TRUE
)
}
\description{
Get rigal trajectory for multiple sites
}
\examples{

tar_load(op_protocol)
tar_load(abun_rich_op)

trends_data <- abun_rich_op \%>\%
 left_join(op_protocol, by = "op_id") \%>\%
 filter(siteid \%in\% mask_siteid)

get_rigal_trajectory_classification(
 trends_data[trends_data$siteid == unique(trends_data$siteid)[2], ],
 y_var = "total_abundance",
 x_var = "year", site_id = "siteid")
}
