% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_spatial_data.R
\name{get_shp_files}
\alias{get_shp_files}
\title{Works with windows shortcuts}
\usage{
get_shp_files(dir = here("inst", "extdata", "RiverATLAS_v10_shp"))
}
\description{
Works with windows shortcuts
}
