% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_raw_data.R
\name{get_site_desc_loc}
\alias{get_site_desc_loc}
\title{Get small tibbles from raw data}
\usage{
get_site_desc_loc(ts_data = NULL, save_data = TRUE)
}
\description{
get_site_desc_loc(): Keep unique values for site location
}
