% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spamm_modelling_utils.R
\name{get_spamm_model}
\alias{get_spamm_model}
\title{Run spamm models}
\usage{
get_spamm_model(
  resp_var = NULL,
  rhs_formula = NULL,
  family = NULL,
  dataset = NULL
)
}
\description{
Run spamm models
}
