% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{get_temporal_vegdist}
\alias{get_temporal_vegdist}
\title{From vegdist object to temporal diversity}
\usage{
get_temporal_vegdist(dist_mat = NULL, drop_first_year = TRUE)
}
\description{
From vegdist object to similarity index from the reference year to the other
}
