% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_distribution.R
\name{get_unique_values}
\alias{get_unique_values}
\title{Get unique values}
\usage{
get_unique_values(x = NULL, na.omit = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{na.omit}{logical.}
}
\value{
vector of length 1 containing the unique value of the vector or the
character "no_unique".
}
\description{
Get unique values
}
\examples{
x <- rep(list(c(runif(1e3), rep(NA, 1000))), 100)
get_unique_values(x)
get_unique_values(x, na.omit = FALSE)
}
