% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model.R
\name{get_untb_abun_mat}
\alias{get_untb_abun_mat}
\title{Wrapper about untb}
\usage{
get_untb_abun_mat(
  start = sample(1:3, size = 64, replace = TRUE),
  prob = 0.001,
  D = 1,
  gens = 1300,
  sparse = TRUE
)
}
\description{
Wrapper about untb
}
