% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_shortcut.R
\name{get_var_localisation}
\alias{get_var_localisation}
\title{Shortcut to obtain vector of variable names}
\usage{
get_var_localisation()
}
\value{
character vector
}
\description{
Shortcut to obtain vector of variable names
}
