% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{get_vegdist_temporal_turnover}
\alias{get_vegdist_temporal_turnover}
\title{Get temporal turnover from first year with vegdist}
\usage{
get_vegdist_temporal_turnover(
  mat = NULL,
  method = NULL,
  return_tibble = TRUE,
  drop_first_year = TRUE
)
}
\description{
Wrapper around vegdist
}
\examples{
tar_load(com_mat_site)
get_vegdist_temporal_turnover(com_mat_site$mat[[1]], method = "jaccard")
}
