% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fishing_operation.R
\name{month_filter}
\alias{month_filter}
\title{Filter operation based on their month}
\usage{
month_filter(x = NULL, range_x = NULL, month_int_decimal = 1)
}
\description{
Filter operation based on their month
}
\examples{
month_filter(x = "2020-01-01", med_x = "2002-01-01", range_x = 1)
month_filter(x = "2020-02-01", med_x = "2002-01-01", range_x = 1)
month_filter(x = "2020-12-03", med_x = "2002-01-01", range_x = 1)
month_filter(x = "2020-11-30", med_x = "2002-01-01", range_x = 1)
# FALSE
month_filter(x = "2020-11-28", med_x = "2002-01-01", range_x = 1)
month_filter(x = "2020-02-15", med_x = "2002-01-01", range_x = 1)
#NA
month_filter(x = NA, med_x = "2002-01-01", range_x = 1)
month_filter(x = "2020-02-30", med_x = "2002-01-01", range_x = 1)
month_filter(x = 1, med_x = "2002-01-01", range_x = 1)
month_filter(x = "2020-02-30", med_x = 1, range_x = 1)
}
