% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spamm_modelling_utils.R
\name{run_save_spamm_model}
\alias{run_save_spamm_model}
\title{Run and save spamm models}
\usage{
run_save_spamm_model(
  resp_var = NULL,
  rhs_formula = NULL,
  family = NULL,
  dataset = NULL
)
}
\description{
Run and save spamm models
}
\examples{
run_save_spamm_model(
 resp_var = "abundance",
 rhs_formula = "year",
 dataset = il_dataset,
 family = "poisson"
)
}
