% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_match.R
\name{st_snap_points}
\alias{st_snap_points}
\title{Span points on geometry}
\usage{
st_snap_points(
  x,
  y,
  max_dist = 1000,
  return_snapped_pt = FALSE,
  river_id = "HYRIV_ID"
)
}
\description{
Span points on geometry
}
\references{
https://github.com/r-spatial/sf/issues/792
}
