% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_distribution.R
\name{summary_distribution}
\alias{summary_distribution}
\title{Get the summary of a numerical vector}
\usage{
summary_distribution(x = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{numerical vector.}

\item{na.rm}{logical.}
}
\value{
a named vector.
}
\description{
compute the "min", "1st_quart", "median", "2nd_quart", "max", "n", number of
na and na fraction.
}
\examples{
x <- c(runif(100), NA)
summary_distribution(x)
summary_distribution(x, na.rm = TRUE)
}
