% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_match.R
\name{target_snap_site_to_river}
\alias{target_snap_site_to_river}
\title{Snap sites on river atlas for every continent}
\usage{
target_snap_site_to_river(
  river_shp_filepath = NULL,
  site_sf = NULL,
  proj_crs = 4087,
  length_chunk = 200,
  max_dist = 1000,
  river_id = "HYRIV_ID"
)
}
\arguments{
\item{proj_crs}{integer default to 4087 because it keeps the distance well globally
https://epsg.org/crs_4087/WGS-84-World-Equidistant-Cylindrical.html
https://epsg.io/4087}
}
\description{
Has parallel capacities with future::plan()
}
