% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{variable_by_year}
\alias{variable_by_year}
\title{Compute by year variable (used for turnover)}
\usage{
variable_by_year(x = NULL, variable = NULL)
}
\description{
Compute by year variable (used for turnover)
}
